Fixed keyboard/mouse scancode handling, added a complete font8x8 table, replaced hlt with pause in main loop to fix freeze

+++ drivers/ps2keyboard.hpp
@@ -1,11 +1,12 @@
 #pragma once
 #include <cstdint>
 
-class PS2Keyboard {
-public:
+class PS2Keyboard
+{
+   public:
     PS2Keyboard() = default;
     void init();
-    int8_t read_byte_nonblocking();
+    int16_t read_byte_nonblocking();
     // Convert set1 scancode to ASCII (very basic US layout)
     static char scancode_to_ascii(uint8_t sc);
 };
